+++ packages/builder/src/components/automation/AutomationBuilder/FlowChart/AutomationStepHelpers.ts
